Moved room creation from /rooms/create route to a modal located at / route
